requirements.txt
asgiref==3.11.0
click==8.3.1
dj-database-url==3.1.0
Django==6.0.1
django-widget-tweaks==1.5.1
gunicorn==25.0.1
h11==0.16.0
packaging==26.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
uvicorn==0.40.0
whitenoise==6.11.0
